Refactor ItemStat component into its own file
